Extract menu in a separate class
