+++ demo/src/main/java/com/teamdev/jxbrowser/demo/ToolBar.java
@@ -21,8 +21,6 @@
 package com.teamdev.jxbrowser.demo;
 
 import com.teamdev.jxbrowser.chromium.Browser;
-import com.teamdev.jxbrowser.chromium.BrowserPreferences;
-import com.teamdev.jxbrowser.chromium.SavePageType;
 import com.teamdev.jxbrowser.chromium.events.FinishLoadingEvent;
 import com.teamdev.jxbrowser.chromium.events.LoadAdapter;
 import com.teamdev.jxbrowser.chromium.events.ProvisionalLoadingEvent;
@@ -32,19 +30,9 @@ import com.teamdev.jxbrowser.chromium.swing.BrowserView;
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionEvent;
-import java.awt.event.InputEvent;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
-import java.io.File;
-
-import static com.teamdev.jxbrowser.demo.resources.Resources.loadIcon;
-import static javax.swing.JOptionPane.INFORMATION_MESSAGE;
-import static javax.swing.JOptionPane.showMessageDialog;
 
 final class ToolBar extends JPanel {
 
-    private static final String RUN_JAVASCRIPT = "Run JavaScript...";
-    private static final String CLOSE_JAVASCRIPT = "Close JavaScript Console";
     private static final String DEFAULT_URL = "about:blank";
 
     private final BrowserView browserView;
@@ -56,6 +44,7 @@ final class ToolBar extends JPanel {
     private JButton forwardButton;
     private JButton refreshButton;
     private JButton stopButton;
+
     private JMenuItem consoleMenuItem;
 
     ToolBar(BrowserView browserView) {
@@ -74,8 +63,10 @@ final class ToolBar extends JPanel {
                 0, 0));
     }
 
+    //TODO:2018-05-08:alexander.yevsyukov: Can we have browser -> browser.goBack() instead?
+
     private static JButton createBackwardButton(final Browser browser) {
-        return createButton("Back", new AbstractAction() {
+        return new ActionButton("Back", new AbstractAction() {
             public void actionPerformed(ActionEvent e) {
                 browser.goBack();
             }
@@ -83,7 +74,7 @@ final class ToolBar extends JPanel {
     }
 
     private static JButton createForwardButton(final Browser browser) {
-        return createButton("Forward", new AbstractAction() {
+        return new ActionButton("Forward", new AbstractAction() {
             public void actionPerformed(ActionEvent e) {
                 browser.goForward();
             }
@@ -91,7 +82,7 @@ final class ToolBar extends JPanel {
     }
 
     private static JButton createRefreshButton(final Browser browser) {
-        return createButton("Refresh", new AbstractAction() {
+        return new ActionButton("Refresh", new AbstractAction() {
             public void actionPerformed(ActionEvent e) {
                 browser.reload();
             }
@@ -99,30 +90,11 @@ final class ToolBar extends JPanel {
     }
 
     private static JButton createStopButton(final Browser browser) {
-        return createButton("Stop", new AbstractAction() {
+        return new ActionButton("Stop", new AbstractAction() {
             public void actionPerformed(ActionEvent e) {
                 browser.stop();
             }
         });
-    }
-
-    private static JButton createButton(String caption, Action action) {
-        ActionButton button = new ActionButton(caption, action);
-        String imageName = caption.toLowerCase();
-        button.setIcon(loadIcon(imageName + ".png"));
-        button.setRolloverIcon(loadIcon(imageName + "-selected.png"));
-        return button;
-    }
-
-    private static JCheckBoxMenuItem createCheckBoxMenuItem(String title, boolean selected,
-            final CheckBoxMenuItemCallback action) {
-        final JCheckBoxMenuItem menuItem = new JCheckBoxMenuItem(title, selected);
-        menuItem.addActionListener(e -> action.call(menuItem.isSelected()));
-        return menuItem;
-    }
-
-    void didJsConsoleClose() {
-        consoleMenuItem.setText(RUN_JAVASCRIPT);
     }
 
     private JPanel createActionsPane() {
@@ -181,246 +153,14 @@ final class ToolBar extends JPanel {
         return result;
     }
 
+    void didJsConsoleClose() {
+        consoleMenuItem.setText(Menu.RUN_JAVASCRIPT);
+    }
+
     private JComponent createMenuButton() {
-        final JPopupMenu popupMenu = new JPopupMenu();
-        popupMenu.add(createConsoleMenuItem());
-        popupMenu.add(createGetHtmlMenuItem());
-        popupMenu.add(createPopupsMenuItem());
-        popupMenu.add(createUploadFileMenuItem());
-        popupMenu.add(createDownloadFileMenuItem());
-        popupMenu.add(createJavaScriptDialogsMenuItem());
-        popupMenu.add(createPDFViewerMenuItem());
-        popupMenu.add(createFlashMenuItem());
-        popupMenu.add(createGoogleMapsMenuItem());
-        popupMenu.add(createHtml5VideoMenuItem());
-        popupMenu.add(createZoomInMenuItem());
-        popupMenu.add(createZoomOutMenuItem());
-        popupMenu.add(createActualSizeMenuItem());
-        popupMenu.add(createSaveWebPageMenuItem());
-        popupMenu.add(createClearCacheMenuItem());
-        popupMenu.add(createPreferencesSubMenu());
-        popupMenu.add(createEditMenu());
-        popupMenu.add(createPrintMenuItem());
-        popupMenu.addSeparator();
-        popupMenu.add(createMoreMenuItem());
-        popupMenu.addSeparator();
-        popupMenu.add(createAboutMenuItem());
-
-        final ActionButton button = new ActionButton("Menu", null);
-        button.setIcon(loadIcon("gear.png"));
-        button.addMouseListener(new MouseAdapter() {
-            public void mousePressed(MouseEvent e) {
-                if ((e.getModifiers() & InputEvent.BUTTON1_MASK) != 0) {
-                    popupMenu.show(e.getComponent(), 0, button.getHeight());
-                } else {
-                    popupMenu.setVisible(false);
-                }
-            }
-        });
-        return button;
-    }
-
-    private Component createPrintMenuItem() {
-        JMenuItem menuItem = new JMenuItem("Print...");
-        menuItem.addActionListener(e -> browser.print());
-        return menuItem;
-    }
-
-    private Component createPreferencesSubMenu() {
-        JMenu menu = new JMenu("Preferences");
-        BrowserPreferences preferences = browser.getPreferences();
-        //TODO:2018-05-08:alexander.yevsyukov: Refactor to have
-        // browser.getPreferences()
-        // browser.setPreferences()
-        // browser.reloadIgnoringCache()
-        // called in the common base.
-        // The callback should only define the "meat" of the call. For example something like:
-        // selected -> { BrowserPreferences::setJavaScriptEnabled(selected) }
-        menu.add(createCheckBoxMenuItem("JavaScript Enabled", preferences.isJavaScriptEnabled(),
-                selected -> {
-                    BrowserPreferences newPrefs = browser.getPreferences();
-                    newPrefs.setJavaScriptEnabled(selected);
-                    browser.setPreferences(newPrefs);
-                    browser.reloadIgnoringCache();
-                }));
-        menu.add(createCheckBoxMenuItem("Images Enabled", preferences.isImagesEnabled(),
-                selected -> {
-                    BrowserPreferences preferences12 = browser.getPreferences();
-                    preferences12.setImagesEnabled(selected);
-                    browser.setPreferences(preferences12);
-                    browser.reloadIgnoringCache();
-                }));
-        menu.add(createCheckBoxMenuItem("Plugins Enabled", preferences.isPluginsEnabled(),
-                selected -> {
-                    BrowserPreferences preferences13 = browser.getPreferences();
-                    preferences13.setPluginsEnabled(selected);
-                    browser.setPreferences(preferences13);
-                    browser.reloadIgnoringCache();
-                }));
-        menu.add(createCheckBoxMenuItem("JavaScript Can Access Clipboard",
-                preferences.isJavaScriptCanAccessClipboard(),
-                selected -> {
-                    BrowserPreferences preferences14 = browser.getPreferences();
-                    preferences14.setJavaScriptCanAccessClipboard(selected);
-                    browser.setPreferences(preferences14);
-                    browser.reloadIgnoringCache();
-                }));
-        return menu;
-    }
-
-    private Component createClearCacheMenuItem() {
-        JMenuItem menuItem = new JMenuItem("Clear Cache");
-        menuItem.addActionListener(
-            e -> browser.getCacheStorage().clearCache(
-                    () -> showMessageDialog(browserView, "Cache is cleared successfully.",
-                            "Clear Cache", INFORMATION_MESSAGE))
-        );
-        return menuItem;
-    }
-
-    private Component createEditMenu() {
-        EditMenu menu = new EditMenu(browserView);
-        return menu.getComponent();
-    }
-
-    private Component createMoreMenuItem() {
-        JMenuItem menuItem = new JMenuItem("More Features...");
-        menuItem.addActionListener(
-                e -> browser.loadURL("https://jxbrowser.support.teamdev.com/support/solutions/9000049010")
-        );
-        return menuItem;
-    }
-
-    private Component createSaveWebPageMenuItem() {
-        JMenuItem menuItem = new JMenuItem("Save Web Page...");
-        menuItem.addActionListener(e -> {
-            JFileChooser fileChooser = new JFileChooser();
-            fileChooser.setSelectedFile(new File("my-web-page.html"));
-            int result = fileChooser.showSaveDialog(browserView);
-            if (result == JFileChooser.APPROVE_OPTION) {
-                File selectedFile = fileChooser.getSelectedFile();
-                String dirPath = new File(selectedFile.getParent(), "resources")
-                        .getAbsolutePath();
-                browser.saveWebPage(selectedFile.getAbsolutePath(),
-                        dirPath,
-                        SavePageType.COMPLETE_HTML);
-            }
-        });
-        return menuItem;
-    }
-
-    private Component createActualSizeMenuItem() {
-        JMenuItem menuItem = new JMenuItem("Actual Size");
-        menuItem.addActionListener(e -> browser.zoomReset());
-        return menuItem;
-    }
-
-    private Component createZoomOutMenuItem() {
-        JMenuItem menuItem = new JMenuItem("Zoom Out");
-        menuItem.addActionListener(e -> browser.zoomOut());
-        return menuItem;
-    }
-
-    private Component createZoomInMenuItem() {
-        JMenuItem menuItem = new JMenuItem("Zoom In");
-        menuItem.addActionListener(e -> browser.zoomIn());
-        return menuItem;
-    }
-
-    private Component createHtml5VideoMenuItem() {
-        JMenuItem menuItem = new JMenuItem("HTML5 Video");
-        menuItem.addActionListener(
-                e -> browser.loadURL("http://www.w3.org/2010/05/video/mediaevents.html")
-        );
-        return menuItem;
-    }
-
-    private Component createGoogleMapsMenuItem() {
-        JMenuItem menuItem = new JMenuItem("Google Maps");
-        menuItem.addActionListener(e -> browser.loadURL("https://maps.google.com/"));
-        return menuItem;
-    }
-
-    private Component createJavaScriptDialogsMenuItem() {
-        JMenuItem menuItem = new JMenuItem("JavaScript Dialogs");
-        menuItem.addActionListener(
-                e -> browser.loadURL("http://www.javascripter.net/faq/alert.htm")
-        );
-        return menuItem;
-    }
-
-    private Component createDownloadFileMenuItem() {
-        JMenuItem menuItem = new JMenuItem("Download File");
-        menuItem.addActionListener(
-                e -> browser.loadURL("https://s3.amazonaws.com/cloud.teamdev.com/downloads/demo/jxbrowserdemo.jnlp")
-        );
-        return menuItem;
-    }
-
-    private Component createGetHtmlMenuItem() {
-        JMenuItem menuItem = new JMenuItem("Get HTML");
-        menuItem.addActionListener(e -> {
-            String html = browser.getHTML();
-            Window window = SwingUtilities.getWindowAncestor(browserView);
-            JDialog dialog = new JDialog(window);
-            dialog.setModal(true);
-            dialog.setContentPane(new JScrollPane(new JTextArea(html)));
-            dialog.setSize(700, 500);
-            dialog.setLocationRelativeTo(null);
-            dialog.setVisible(true);
-
-        });
-        return menuItem;
-    }
-
-    private JMenuItem createConsoleMenuItem() {
-        consoleMenuItem = new JMenuItem(RUN_JAVASCRIPT);
-        consoleMenuItem.addActionListener(e -> {
-            if (RUN_JAVASCRIPT.equals(consoleMenuItem.getText())) {
-                consoleMenuItem.setText(CLOSE_JAVASCRIPT);
-                firePropertyChange("JSConsoleDisplayed", false, true);
-            } else {
-                consoleMenuItem.setText(RUN_JAVASCRIPT);
-                firePropertyChange("JSConsoleClosed", false, true);
-            }
-        });
-        return consoleMenuItem;
-    }
-
-    private JMenuItem createUploadFileMenuItem() {
-        JMenuItem menuItem = new JMenuItem("Upload File");
-        menuItem.addActionListener(
-                e -> browser.loadURL("http://www.cs.tut.fi/~jkorpela/forms/file.html#example")
-        );
-        return menuItem;
-    }
-
-    private JMenuItem createPopupsMenuItem() {
-        JMenuItem menuItem = new JMenuItem("Popup Windows");
-        menuItem.addActionListener(e -> browser.loadURL("http://www.popuptest.com"));
-        return menuItem;
-    }
-
-    private JMenuItem createPDFViewerMenuItem() {
-        JMenuItem menuItem = new JMenuItem("PDF Viewer");
-        menuItem.addActionListener(e -> browser.loadURL("http://www.orimi.com/pdf-test.pdf"));
-        return menuItem;
-    }
-
-    private JMenuItem createFlashMenuItem() {
-        JMenuItem menuItem = new JMenuItem("Adobe Flash");
-        menuItem.addActionListener(e -> browser.loadURL("http://helpx.adobe.com/flash-player.html"));
-        return menuItem;
-    }
-
-    private JMenuItem createAboutMenuItem() {
-        JMenuItem menuItem = new JMenuItem("About JxBrowser Demo");
-        menuItem.addActionListener(e -> {
-            Frame parentFrame = (Frame) SwingUtilities.getWindowAncestor(ToolBar.this);
-            AboutDialog aboutDialog = new AboutDialog(parentFrame);
-            aboutDialog.setVisible(true);
-        });
-        return menuItem;
+        Menu menu = new Menu(this, browserView);
+        consoleMenuItem = menu.getConsoleMenuItem();
+        return menu.getButton();
     }
 
     private boolean isFocusRequired() {
@@ -437,23 +177,5 @@ final class ToolBar extends JPanel {
                 addressBar.selectAll();
             }
         });
-    }
-
-    private interface CheckBoxMenuItemCallback {
-        void call(boolean selected);
-    }
-
-    private static class ActionButton extends JButton {
-        private ActionButton(String hint, Action action) {
-            super(action);
-            setContentAreaFilled(false);
-            setBorder(BorderFactory.createEmptyBorder());
-            setBorderPainted(false);
-            setRolloverEnabled(true);
-            setToolTipText(hint);
-            setText(null);
-            setFocusable(false);
-            setDefaultCapable(false);
-        }
     }
 }
