+++ demo/src/main/resources/com/teamdev/jxbrowser/demo/resources/icon/menu-selected.png
